requirements.txt
pyspark==3.5.0
requests==2.32.3
python-dotenv==1.1.0
pytest==8.3.5
ruff
